Pull in a test the new `yk_indirect_inline` attribute.
This exposes a deficiency in the jitc_yk trace builder, which I've
hacked around for now.

We will get an assertion failure if/when this deficiency arises, but
hopefully j2 has a way to handle it properly.
